refactor: 기존 구현 코드 RacingCars 의존성 주입 적용
- DB 엔티티와 도메인 객체 간 일관성을 위한 리팩터링 진행중
- DB 조회 정보로 RacingCars 객체 생성 가능하게 하기 위함
