+++ src/main/java/racingcar/controller/RacingCarController.java
@@ -6,6 +6,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+import racingcar.domain.RacingCarNames;
 import racingcar.dto.PlayRequestDto;
 import racingcar.dto.PlayResponseDto;
 import racingcar.service.RacingCarService;
@@ -26,7 +27,7 @@ public class RacingCarController {
         final int count = playRequestDto.getCount();
         final List<String> carNames = TextParser.parseByDelimiter(playRequestDto.getNames(), CAR_NAMES_DELIMITER);
 
-        final PlayResponseDto playResult = racingCarService.playGame(count, carNames);
+        final PlayResponseDto playResult = racingCarService.playGame(count, new RacingCarNames(carNames));
 
         return ResponseEntity.ok(playResult);
     }
